fix trunc_v1

+++ src/Decimal/4_Anouther_function/s21_truncate.c
@@ -4,7 +4,7 @@ int s21_truncate(s21_decimal value, s21_decimal *result) {
   int flag = 0;  //// добавить ошибки при вычислении (если переменные не по
                  /// формату и тп)
   s21_DecData val = s21_decimal_get_data(value);
-  if (val.high_bit - 3 * val.scale >= 0) {
+  if ((int)val.high_bit - 3 * val.scale >= 0) {
     for (int i = val.scale; i > 0; i--) {
       s21_div_10(&val);
     }
